% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nested_value}
\alias{get_nested_value}
\title{Get a nested value from a JSON object using a path}
\usage{
get_nested_value(json_data, path, default = NULL)
}
\arguments{
\item{json_data}{A JSON object (as an R list)}

\item{path}{Path to the value using dot notation (e.g., "person.address.city")}

\item{default}{Default value to return if the path does not exist}
}
\value{
The value at the specified path or the default value if the path does not exist
}
\description{
Get a nested value from a JSON object using a path
}
\keyword{internal}
