% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_json}
\alias{is_valid_json}
\title{Check if a JSON string or file is valid}
\usage{
is_valid_json(json_input, is_file = FALSE)
}
\arguments{
\item{json_input}{Either a JSON string or a path to a JSON file}

\item{is_file}{Logical indicating whether the input is a file path}
}
\value{
Logical indicating whether the JSON is valid
}
\description{
This function checks if a JSON string or file is valid according to JSON syntax.
}
\keyword{internal}
