% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pretty_json}
\alias{pretty_json}
\title{Format JSON for pretty printing}
\usage{
pretty_json(json_input)
}
\arguments{
\item{json_input}{Either a JSON object (as an R list) or a JSON string}
}
\value{
A formatted JSON string
}
\description{
This function formats a JSON object or string for pretty printing.
}
\keyword{internal}
