% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonr-package.R
\docType{package}
\name{jsonr-package}
\alias{jsonr-package}
\alias{jsonr}
\title{jsonr: Advanced JSON Manipulation Tools for R}
\description{
Advanced tools for manipulating JSON data in R. The package provides functions for adding elements to specific paths in JSON objects, inserting properties into nested array elements, and other JSON manipulation tasks. It offers a more flexible and powerful approach to working with complex JSON structures than what is available in base packages.

The \code{jsonr} package provides tools for advanced manipulation of JSON data in R.
It offers functions for adding elements to specific paths in JSON objects, inserting
properties into nested array elements, and other JSON manipulation tasks.
}
\section{Main Functions}{

\itemize{
\item \code{\link{add_json_element}}: Add an element to a structure in a JSON object or file
\item \code{\link{insert_json_property}}: Insert a property into elements of a nested JSON array
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luechtian/jsonr}
  \item Report bugs at \url{https://github.com/luechtian/jsonr/issues}
}

}
\author{
\strong{Maintainer}: Christian Luechtenborg \email{christian.luechtenborg@gmail.com}

}
\keyword{internal}
